// @cloud-materials/common/components/CSearch/interface.ts
export interface PropsWithBase {
  /** 类名 */
  className?: string | string[];
  /** 样式 */
  style?: CSSProperties;
}

/** 组件值基类 */
export type PropsWithValue<T> = {
  /** 当前值（受控） */
  value?: T;
  /** 默认值（非受控） */
  defaultValue?: T;
  /** 值改变事件 */
  onChange?: (value: T) => void;
};

export interface PropsWithDebounce {
  /**
   * @zh 防抖参数
   * @defaultValue `{wait: 500}`
   */
  debounceOptions?:
    | (DebounceSettings & {
        wait?: number;
      })
    | null;
}

/** 组件类型 */
export type { CSearchComponentType };
/** 搜索参数 */
export type CSearchParams = Record<string, any>;
/** 组件内容 */
export type CSearchComponentContent =
  | (ReactElement & { component?: undefined })
  | DefineBuiltInBasicType<CSearchComponentEnum>;

/** 简单搜索组件参数 */
export interface CSearchComponentProps {
  /** 内容 */
  content: CSearchComponentContent;
  /** 组件参数 */
  commonProps?: Record<string, any>;
}
// ...
export interface CCascaderSearchItem extends CSearchComponentProps {
  /** 选项名称 */
  label: ReactNode;
  /** 选项字段 */
  fieldName: Key;
  /**
   * @zh 是否可见
   * @defaultValue true
   */
  visible?: boolean;
}

/**
 * @title CCascaderSearch
 */
export interface CCascaderSearchProps extends PropsWithBase, PropsWithDebounce, PropsWithValue<CSearchParams> {
  /** 搜索类型选择项 */
  list?: CCascaderSearchItem[];
  /** 搜索类型选择器参数 */
  arcoSelectProps?: Omit<SelectProps, 'options' | 'value'>;
  /**
   * @zh 标签宽度
   * @defaultValue 100
   */
  labelWidth?: CSSProperties['width'];
  /**
   * @zh 搜索项Select组件是否配置自动宽度，开启后，labelWidth失效
   * @defaultValue false
   */
  prefixAutoWidth?: boolean;
}

/**
 * @title UseCCascaderSearchValue
 */
export interface UseCCascaderSearchValue {
  /** 过滤掉 visible 为 false 的选项 */
  validOptions: CCascaderSearchItem[];
  /** 当前值 */
  state: {
    /** 字段 */
    fieldName: Key;
    /** 值 */
    value?: any;
  };
  /** 组件 */
  searchComponent: CSearchComponentProps;
}

/**
 * @title UseCCascaderSearchControl
 */
export interface UseCCascaderSearchControl {
  /** 更新字段 */
  updateField: (value: Key) => void;
  /** 更新值 */
  updateValue: (value: any) => void;
}